fixed rng for heuristic 1

+++ src/first_alg_with_heuristics.h
@@ -1,3 +1,4 @@
+#include <iostream>
 #include <cmath>
 #include <vector>
 #include <cstdint>
@@ -9,7 +10,7 @@
 
 std::function<int(size_t)> generate_hash(size_t num_buckets);
 
-std::function<int(size_t)> choose_hash_heuristic_1(size_t num_buckets, int k, std::map<int, int>& freq);
+std::function<int(size_t)> choose_hash_heuristic_1(size_t num_buckets, int k, std::map<int, long long>& freq);
 
 void HammingDistanceBase(size_t n, size_t m, size_t sigma, const std::vector<uint32_t>& A, const std::vector<uint32_t>& B, std::vector<uint32_t>& dist, std::mt19937& rng);
